gcdll-back/requirements_dev.txt
-r requirements.txt

coverage
pytest

pycodestyle
pylint
pylint-django